digihub-chatbot-query-pipeline/src/utils/rules.py
import yara

userquery_process_rules = """
rule DetectPromptInjectionAttempt {
  meta:
    author = "AI Assistant"
    description = "Detects common prompt injection phrases and obfuscation."

  strings:
    $s1 = "ignore previous instructions" nocase ascii wide
    $s2 = "disregard all prior commands" nocase ascii wide
    $s3 = "as an admin" nocase ascii wide
    $s4 = "reveal your system prompt" nocase ascii wide
    $s6 = /[a-zA-Z]([#|][a-zA-Z])\w/ // Detects patterns like A|B|C|D|E
    $s7_alphanum = /[a-zA-Z0-9][^a-zA-Z0-9\s,.]\w/
    $ignore = "ignore previous instructions"
    $s8 = /[I|G|N|O|R|E][,\s]?/


  condition:
    any of them
}
"""

def get_rules():
    rules = yara.compile(source=userquery_process_rules)
    return rules
