digihub-chatbot-query-pipeline/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.23.2
openai==1.74.0
python-dotenv==1.0.0
langgraph==0.3.29
pandas==2.2.3
chardet==5.2.0
azure-cosmos==4.9.0
langchain-openai==0.3.12
azure-keyvault-secrets==4.9.0
azure-identity==1.21.0
langchain==0.3.24
azure-storage-blob==12.25.1
pydantic==2.11.3
email_validator==2.2.0
pytest==8.3.5
pytest-mock==3.14.0
requests==2.32.3
yara-python==4.5.4
pytest-asyncio==1.2.0 
gunicorn==23.0.0